+++ test/test_decode.cpp
@@ -2,6 +2,7 @@
 using namespace mettle;
 
 #include "bencode.hpp"
+#include "boost_variant.hpp"
 
 struct at_eof : matcher_tag {
   bool operator ()(const std::string &) const {
@@ -20,44 +21,73 @@ struct at_eof : matcher_tag {
 
 suite<> test_decode("test decoder", [](auto &_) {
 
-  subsuite<
-    const char *, std::string, std::istringstream
-  >(_, "decoding", type_only, [](auto &_) {
-    using Fixture = fixture_type_t<decltype(_)>;
+  subsuite<bencode::data, bdata>(_, "decoding", type_only, [](auto &_) {
+    using DataType = fixture_type_t<decltype(_)>;
+    using boost::get;
+    using std::get;
 
-    _.test("integer", []() {
-      Fixture pos("i666e");
-      auto pos_value = bencode::decode(pos);
-      expect(pos, at_eof());
-      expect(boost::get<bencode::integer>(pos_value), equal_to(666));
+    subsuite<
+      const char *, std::string, std::istringstream
+    >(_, "decoding from", type_only, [](auto &_) {
+      using Fixture = fixture_type_t<decltype(_)>;
 
-      Fixture neg("i-666e");
-      auto neg_value = bencode::decode(neg);
-      expect(neg, at_eof());
-      expect(boost::get<bencode::integer>(neg_value), equal_to(-666));
-    });
+      _.test("integer", []() {
+        Fixture pos("i666e");
+        auto pos_value = bencode::basic_decode<DataType>(pos);
+        expect(pos, at_eof());
+        expect(get<typename DataType::integer>(pos_value), equal_to(666));
 
-    _.test("string", []() {
-      Fixture data("4:spam");
-      auto value = bencode::decode(data);
-      expect(data, at_eof());
-      expect(boost::get<bencode::string>(value), equal_to("spam"));
-    });
+        Fixture neg("i-666e");
+        auto neg_value = bencode::basic_decode<DataType>(neg);
+        expect(neg, at_eof());
+        expect(get<typename DataType::integer>(neg_value), equal_to(-666));
+      });
 
-    _.test("list", []() {
-      Fixture data("li666ee");
-      auto value = bencode::decode(data);
-      auto list = boost::get<bencode::list>(value);
-      expect(data, at_eof());
-      expect(boost::get<bencode::integer>(list[0]), equal_to(666));
-    });
+      _.test("string", []() {
+        Fixture data("4:spam");
+        auto value = bencode::basic_decode<DataType>(data);
+        expect(data, at_eof());
+        expect(get<typename DataType::string>(value), equal_to("spam"));
+      });
 
-    _.test("dict", []() {
-      Fixture data("d4:spami666ee");
-      auto value = bencode::decode(data);
-      auto dict = boost::get<bencode::dict>(value);
-      expect(data, at_eof());
-      expect(boost::get<bencode::integer>(dict["spam"]), equal_to(666));
+      _.test("list", []() {
+        Fixture data("li666ee");
+        auto value = bencode::basic_decode<DataType>(data);
+        auto list = get<typename DataType::list>(value);
+        expect(data, at_eof());
+        expect(get<typename DataType::integer>(list[0]), equal_to(666));
+      });
+
+      _.test("dict", []() {
+        Fixture data("d4:spami666ee");
+        auto value = bencode::basic_decode<DataType>(data);
+        auto dict = get<typename DataType::dict>(value);
+        expect(data, at_eof());
+        expect(get<typename DataType::integer>(dict["spam"]), equal_to(666));
+      });
+
+      _.test("nested", []() {
+        Fixture data("d"
+          "3:one" "i1e"
+          "5:three" "l" "d" "3:bar" "i0e" "3:foo" "i0e" "e" "e"
+          "3:two" "l" "i3e" "3:foo" "i4e" "e"
+        "e");
+
+        auto value = bencode::basic_decode<DataType>(data);
+        expect(data, at_eof());
+        auto dict = get<typename DataType::dict>(value);
+        expect(get<typename DataType::integer>(dict["one"]), equal_to(1));
+
+        expect(get<typename DataType::string>(
+          get<typename DataType::list>(dict["two"])[1]
+        ), equal_to("foo"));
+
+        expect(get<typename DataType::integer>(
+          get<typename DataType::dict>(
+            get<typename DataType::list>(dict["three"])[0]
+          )["foo"]
+        ), equal_to(0));
+      });
     });
   });
 
@@ -67,34 +97,40 @@ suite<> test_decode("test decoder", [](auto &_) {
       auto begin = data.begin(), end = data.end();
 
       auto first = bencode::decode(begin, end);
-      expect(boost::get<bencode::integer>(first), equal_to(666));
+      expect(std::get<bencode::integer>(first), equal_to(666));
 
       auto second = bencode::decode(begin, end);
-      expect(boost::get<bencode::string>(second), equal_to("goat"));
+      expect(std::get<bencode::string>(second), equal_to("goat"));
     });
 
     _.test("from stream", []() {
       std::stringstream data("i666e4:goat");
 
       auto first = bencode::decode(data);
-      expect(boost::get<bencode::integer>(first), equal_to(666));
+      expect(std::get<bencode::integer>(first), equal_to(666));
       expect(data, is_not(at_eof()));
 
       auto second = bencode::decode(data);
-      expect(boost::get<bencode::string>(second), equal_to("goat"));
+      expect(std::get<bencode::string>(second), equal_to("goat"));
       expect(data, at_eof());
     });
   });
 
-  subsuite<>(_, "decode_view", [](auto &_) {
+  subsuite<
+    bencode::data_view, bdata_view
+  >(_, "decoding (view)", type_only, [](auto &_) {
+    using DataType = fixture_type_t<decltype(_)>;
+    using boost::get;
+    using std::get;
+
     _.test("integer", []() {
       std::string pos("i666e");
-      auto pos_value = bencode::decode_view(pos);
-      expect(boost::get<bencode::integer_view>(pos_value), equal_to(666));
+      auto pos_value = bencode::basic_decode<DataType>(pos);
+      expect(get<typename DataType::integer>(pos_value), equal_to(666));
 
       std::string neg("i-666e");
-      auto neg_value = bencode::decode_view(neg);
-      expect(boost::get<bencode::integer_view>(neg_value), equal_to(-666));
+      auto neg_value = bencode::basic_decode<DataType>(neg);
+      expect(get<typename DataType::integer>(neg_value), equal_to(-666));
     });
 
     _.test("string", []() {
@@ -104,8 +140,8 @@ suite<> test_decode("test decoder", [](auto &_) {
         less_equal(data.data() + data.size())
       );
 
-      auto value = bencode::decode_view(data);
-      auto str = boost::get<bencode::string_view>(value);
+      auto value = bencode::basic_decode<DataType>(data);
+      auto str = get<typename DataType::string>(value);
       expect(&*str.begin(), in_range);
       expect(&*str.end(), in_range);
       expect(str, equal_to("spam"));
@@ -113,9 +149,9 @@ suite<> test_decode("test decoder", [](auto &_) {
 
     _.test("list", []() {
       std::string data("li666ee");
-      auto value = bencode::decode_view(data);
-      auto list = boost::get<bencode::list_view>(value);
-      expect(boost::get<bencode::integer_view>(list[0]), equal_to(666));
+      auto value = bencode::basic_decode<DataType>(data);
+      auto list = get<typename DataType::list>(value);
+      expect(get<typename DataType::integer>(list[0]), equal_to(666));
     });
 
     _.test("dict", []() {
@@ -125,14 +161,36 @@ suite<> test_decode("test decoder", [](auto &_) {
         less_equal(data.data() + data.size())
       );
 
-      auto value = bencode::decode_view(data);
-      auto dict = boost::get<bencode::dict_view>(value);
+      auto value = bencode::basic_decode<DataType>(data);
+      auto dict = get<typename DataType::dict>(value);
       auto str = dict.find("spam")->first;
       expect(&*str.begin(), in_range);
       expect(&*str.end(), in_range);
       expect(str, equal_to("spam"));
 
-      expect(boost::get<bencode::integer_view>(dict["spam"]), equal_to(666));
+      expect(get<typename DataType::integer>(dict["spam"]), equal_to(666));
+    });
+
+    _.test("nested", []() {
+      std::string data("d"
+        "3:one" "i1e"
+        "5:three" "l" "d" "3:bar" "i0e" "3:foo" "i0e" "e" "e"
+        "3:two" "l" "i3e" "3:foo" "i4e" "e"
+      "e");
+
+      auto value = bencode::basic_decode<DataType>(data);
+      auto dict = get<typename DataType::dict>(value);
+      expect(get<typename DataType::integer>(dict["one"]), equal_to(1));
+
+      expect(get<typename DataType::string>(
+        get<typename DataType::list>(dict["two"])[1]
+      ), equal_to("foo"));
+
+      expect(get<typename DataType::integer>(
+        get<typename DataType::dict>(
+          get<typename DataType::list>(dict["three"])[0]
+        )["foo"]
+      ), equal_to(0));
     });
   });