+++ CHANGES.md
@@ -2,7 +2,11 @@
 
 ## v0.3 (in progress)
 
-- Improve performance of `decode`
+- Replace usage of `boost::variant` with `std::variant` (thus requiring C++17 or
+  newer)
+- Allow customizing the variant type used via `bencode::basic_data`
+- Improve performance of `decode`; decoding is now ~2x as fast for most data
+  (~1.5x when using views)!
 
 ## v0.2.1 (2020-12-03)